Current

+++ images/03.ServiceImages/logview/home/engines/scripts/services/rm_service.sh
@@ -7,5 +7,9 @@ required_values="log_name"
 check_required_values
 log_name=`echo $log_name | sed "/ /s//_/g"`
 config_file=/home/app/config.user.d/${container_type}s/$parent_engine/$log_name.json
-rm $config_file
+#file might not exist if it was rejected due to missing log file
+if test -f $config_file
+  then
+	rm $config_file
+fi	
 #/home/engines/scripts/services/build_config.sh
